ebook sharding demo

+++ ebook-sharding-demo/src/main/java/com/redocon/shardingjdbc/entity/Library.java
@@ -1,24 +1,24 @@
 package com.redocon.shardingjdbc.entity;
 
 public class Library {
-    private Integer id;
-    private Integer customerId;
+    private Long id;
+    private Long customerId;
     private String name;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
 
-    public Integer getCustomerId() {
+    public Long getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(Integer customerId) {
+    public void setCustomerId(Long customerId) {
         this.customerId = customerId;
     }
