+++ ebook-sharding-demo/src/main/java/com/redocon/shardingjdbc/entity/Order.java
@@ -2,23 +2,23 @@ package com.redocon.shardingjdbc.entity;
 
 public class Order {
 
-    private Integer id;
-    private Integer customerId;
+    private Long id;
+    private Long customerId;
     private String name;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCustomerId() {
+    public Long getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(Integer customerId) {
+    public void setCustomerId(Long customerId) {
         this.customerId = customerId;
     }
